Move google auth common logic to the oidc auth executor
