+++ backend/internal/executor/googleauth/constants.go
@@ -25,6 +25,3 @@ const (
 	googleUserInfoEndpoint  = "https://openidconnect.googleapis.com/v1/userinfo"
 	googleJwksEndpoint      = "https://www.googleapis.com/oauth2/v3/certs"
 )
-
-// idTokenNonUserAttributes contains the list of non-user attributes that are expected in the ID token.
-var idTokenNonUserAttributes = []string{"aud", "exp", "iat", "iss", "at_hash", "azp", "nonce", "sub"}
